Move first batch of property tests from inside UnusedPrivateMemberSpec
